rails 7'ye yukseltildi

+++ db/schema.rb
@@ -2,8 +2,8 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
@@ -14,8 +14,8 @@ ActiveRecord::Schema.define(version: 2020_01_05_103459) do
 
   create_table "comments", force: :cascade do |t|
     t.text "comment"
-    t.integer "user_id", null: false
-    t.integer "sound_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "sound_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["sound_id"], name: "index_comments_on_sound_id"
@@ -23,14 +23,14 @@ ActiveRecord::Schema.define(version: 2020_01_05_103459) do
   end
 
   create_table "order_items", force: :cascade do |t|
-    t.integer "order_id", null: false
-    t.integer "sound_id", null: false
+    t.bigint "order_id", null: false
+    t.bigint "sound_id", null: false
     t.index ["order_id"], name: "index_order_items_on_order_id"
     t.index ["sound_id"], name: "index_order_items_on_sound_id"
   end
 
   create_table "orders", force: :cascade do |t|
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.float "total_price"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
@@ -42,7 +42,7 @@ ActiveRecord::Schema.define(version: 2020_01_05_103459) do
     t.string "tag"
     t.string "location"
     t.float "price"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["user_id"], name: "index_sounds_on_user_id"
